refactor(metrics): address review feedback on time-range flags
Refactoring (DRY):
- Extract parseTimeFlagAs shared core; parseTimeFlag and parseTimeFlagMs
  now delegate to it, eliminating ~25 lines of duplication.
- Extract addTimeRangeFlagsMs helper that registers --start/--end and
  returns a closure to parse them; used by all three metrics subcommands.
- Drop redundant 'client' import alias in metrics.go for consistency
  with other files in internal/cmd.
- Convert parseTimeFlag/parseTimeFlagMs tests to a table-driven layout
  parameterized on (parser, unit) so cases stay in sync.

Correctness:
- Mark start/end query schemas as format: int64 in api/openapi.yaml;
  regenerate the client so ListNamesParams, GetLabelsByIdParams, and
  GetValuesByIdParams use int64 instead of int (avoids wraparound on
  32-bit GOARCH where epoch ms overflows int).
- Drop the int(start)/int(end) narrowing casts in metrics.go.
- Update flag descriptions to 'epoch milliseconds' (matches the parser's
  error message wording).

Tests:
- Add request-inspecting mock tests for all three metrics subcommands
  that assert start/end query values land in the wire request.
- Extract metricsQueryServer helper to share that scaffolding.
- Add a TODO note explaining why TestMetricsNames is skipped server-side.

+++ internal/client/types.gen.go
@@ -1321,28 +1321,28 @@ type OodleUtilHttputilsModelsErrors struct {
 // ListNamesParams defines parameters for ListNames.
 type ListNamesParams struct {
 	// Start Start of the time range in epoch milliseconds
-	Start int `form:"start" json:"start"`
+	Start int64 `form:"start" json:"start"`
 
 	// End End of the time range in epoch milliseconds
-	End int `form:"end" json:"end"`
+	End int64 `form:"end" json:"end"`
 }
 
 // GetLabelsByIdParams defines parameters for GetLabelsById.
 type GetLabelsByIdParams struct {
 	// Start Start of the time range in epoch milliseconds
-	Start int `form:"start" json:"start"`
+	Start int64 `form:"start" json:"start"`
 
 	// End End of the time range in epoch milliseconds
-	End int `form:"end" json:"end"`
+	End int64 `form:"end" json:"end"`
 }
 
 // GetValuesByIdParams defines parameters for GetValuesById.
 type GetValuesByIdParams struct {
 	// Start Start of the time range in epoch milliseconds
-	Start int `form:"start" json:"start"`
+	Start int64 `form:"start" json:"start"`
 
 	// End End of the time range in epoch milliseconds
-	End int `form:"end" json:"end"`
+	End int64 `form:"end" json:"end"`
 }
 
 // GetMonitorStateByIdParams defines parameters for GetMonitorStateById.
